requirements.txt
pandas==1.2.1
matplotlib==3.3.4
MetaTrader5==5.0.34
pytz==2021.1
scipy==1.6.0
logging==0.4.9.6
pyyaml==5.4.1
wxpython==4.1.1